backend/requirements.txt
flask==2.3.2
flask-cors==4.0.0
flask-sqlalchemy==3.0.3
psycopg2-binary==2.9.7
python-dotenv==1.0.0
pyjwt==2.7.0
bcrypt==4.0.1